cleaned up files, converted gene_drive script to a All-In-One component, updated app.py
