+++ Gene_Drive/components/footer.py
@@ -5,7 +5,7 @@ import datetime
 current_year = datetime.date.today().year
 
 footer_style = {
-    "position": "fixed",
+    "position": "relative",
     "left": 0,
     "bottom": 0,
     "width": "100%",
