+++ Gene_Drive/requirements.txt
@@ -12,23 +12,23 @@ click==8.0.1
     #   pip-tools
 colorama==0.4.4
     # via click
-dash==1.21.0
+dash==2.0.0rc2
     # via
     #   -r requirements.in
     #   dash-bootstrap-components
 dash-bootstrap-components==0.13.0
     # via -r requirements.in
-dash-core-components==1.17.1
+dash-core-components==2.0.0rc1
     # via
     #   -r requirements.in
     #   dash
-dash-html-components==1.1.4
+dash-html-components==2.0.0rc1
     # via
     #   -r requirements.in
     #   dash
 dash-renderer==1.9.1
     # via -r requirements.in
-dash-table==4.12.0
+dash_table==5.0.0rc1
     # via dash
 flask==2.0.1
     # via
